+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 #### Network
 - Added adaptiveLR as a network configuration
 #### NetMath
+- Added adam as adaptiveLR configuration
 - Added RMSProp as adaptiveLR configuration
-- Added Adagrad as adaptiveLR configuration
+- Added adagrad as adaptiveLR configuration
 - Added gain as adaptiveLR configuration
 - Added Mean Squared Error cost function
 # v1.0.0